backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.4.0
python-decouple==3.8
xhtml2pdf==0.2.15
WeasyPrint==61.0
django-filter==24.2
requests==2.32.3
black==24.4.2
flake8==7.1.1
Pillow==10.3.0
#psycopg2-binary==2.9.9